Add FitsTool script. And, Fixes #18
